Closes #9. Marked parallel sampling services as closeable.
The AnchorConstruction should not be concerned with opening or closing the ExecutorService. After all, the user is required to pass an instance of a generically defined interface and should thus also be responsible for maintaining its state. Closing it automatically after each explanation would require the implementation to reopen it since otherwise the builder would not be usable any more.